requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
nltk>=3.8.1
psutil>=5.9.0
Pillow>=10.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
numpy>=1.24.0
opencv-python>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0 